Move executables in order to be able to run entire test suite
