--- back-end/flush.go
@@ -1,13 +1,0 @@
-package main
-
-import "github.com/go-redis/redis"
-
-func main() {
-	redisClient := redis.NewClient(&redis.Options{
-		Addr:     "localhost:6379",
-		Password: "", // no password set
-		DB:       0,  // use default DB
-	})
-
-	redisClient.FlushAll()
-}
