+++ back-end/new_world/new_world.go
@@ -1,8 +1,8 @@
-package main
+package new_world
 
 import (
-	"./rc"
-	"./wo"
+	"../rc"
+	"../wo"
 	"fmt"
 )
