--- back-end/websocket_server.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"./ws"
-)
-
-func main() {
-	ws.RunServer()
-}
