+++ back-end/websocket_server/websocket_server.go
@@ -1,0 +1,7 @@
+package websocket_server
+
+import "../ws"
+
+func main() {
+	ws.RunServer()
+}
